\name{computeSSDs}
\alias{computeSSDs}
\title{
    Computes possible SSDs for upcoming trial.
}
\description{
    Computes possible SSDs for upcoming trial, based on current predicted RT, threshold parameters and SSD resolution.
}
\usage{
    new_ssds <- computeSSDs( PSIobject )
}
\arguments{
    \item{PSIobject}{a PSI object containing thresholds, predicted RT and SSD resolution}
}
\value{
returns new ssds
}

        