\name{createPrior}
\alias{createPrior}
\title{
    Creates initial prior distribution.
}
\description{
    Creates initial prior distribution (uniform).
}
\usage{
    priors <- createPrior( PSIobject )
}
\arguments{
    \item{PSIobject}{a PSI object containing thresholds, slopes, error_rates and number of Stop-trials}
}
\value{
returns uniform priors
}

        