\name{posteriorSuccess}
\alias{posteriorSuccess}
\title{
    Compute posterior distribution in case of success
}
\description{
    Compute posterior distribution in case of success (inhibit).
}
\usage{
    posterior_success <- posteriorSuccess( PSIobject )
}
\arguments{
    \item{PSIobject}{a PSI object containing updated thresholds, slopes, error_rates, updated SSDs, prior distribution and p(Success) lookup table}
}
\value{
returns posterior distribution for no-error response
}

        